Q1_MATH/requirements.txt
openai==0.28.1
numpy>=1.21.0
python-dotenv>=0.19.0
pathlib2>=2.3.7
requests>=2.25.1